+++ runtime/common/src/tests/setup/mod.rs
@@ -25,7 +25,7 @@ pub use sp_std::cell::RefCell;
 
 use primitives::{AccountId, Balance, BlockNumber};
 
-use crate::{currency::UNIT, xcm_impl::BaseRuntimeRequirements};
+use crate::{currency::UNIT, BaseRuntimeRequirements};
 
 pub const ALICE: [u8; 32] = [1u8; 32];
 pub const BOB: [u8; 32] = [2u8; 32];
