+++ runtime/litentry/src/lib.rs
@@ -688,6 +688,9 @@ impl pallet_extrinsic_filter::Config for Runtime {
 	type WeightInfo = weights::pallet_extrinsic_filter::WeightInfo<Runtime>;
 }
 
+#[cfg(test)]
+impl runtime_common::BaseRuntimeRequirements for Runtime {}
+
 construct_runtime! {
 	pub enum Runtime where
 		Block = Block,
